Update garm-provider-common
This version of garm-provider-common contains a fix for Windows userdata
which now creates a dedicated user under which the github runner is started
as a service. This new user has SeLoginRight and is part of the local
Administrators group. Some applications and workflows require an actual
user with a proper user profile. This change should allow any workflow
to function properly on Windows.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>